requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
better-profanity==0.7.0
blinker==1.9.0
bs4==0.0.2
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
coverage==7.10.1
cryptography==45.0.3
Deprecated==1.2.18
dotenv==0.9.9
firebase==4.0.1
firebase-admin==6.8.0
Flask==3.1.0
Flask-Dance==7.1.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-talisman==1.1.0
Flask-SocketIO==5.5.1
frozenlist==1.7.0
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.173.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-run==0.10.18
google-cloud-secret-manager==2.24.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
ImageHash==4.3.2
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.6.3
numpy==2.3.1
oauthlib==3.2.2
optional==0.0.1
ordered-set==4.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dotenv==1.1.0
PyWavelets==1.8.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9.1
scipy==1.16.0
soupsieve==2.7
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
URLObject==2.4.3
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
zipp==3.21.0
